Rename tests projects. Refactor file-writing cleanup
