+++ database-component-tests/src/main/java/database/component/tests/MotorCycleBuilder.java
@@ -1,4 +1,4 @@
-package database.component.test;
+package database.component.tests;
 
 public class MotorCycleBuilder {
