+++ database-component-tests/src/main/java/database/component/tests/MotorCycleVerifier.java
@@ -1,4 +1,4 @@
-package database.component.test;
+package database.component.tests;
 
 import static org.junit.Assert.*;
