+++ database-service/src/main/java/database/service/factory/DatabaseClientAbstractFactory.java
@@ -21,4 +21,5 @@ public abstract class DatabaseClientAbstractFactory<T extends DatabaseClient> {
   public abstract DatabaseClient getProxiedClient();
 
   protected abstract T instantiate(String databaseName) throws DatabaseLoadingException, DeserializationException;
+
 }
