+++ serialization-manager-component-tests/src/main/java/serialization/manager/component/test/ObjectWithoutId.java
