+++ serialization-manager-component-tests/src/test/java/serialization/manager/component/test/SerializationManagerTest.java
